requirements.txt
streamlit
matplotlib
librosa
wordcloud
jieba
Pillow
numpy
